manim-mcp/server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Manim MCP Server - Educational math visualization tools via Model Context Protocol"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core MCP and web framework
    "fastmcp>=2.12.4,<3",
    "uvicorn[standard]>=0.30.0",
    # Rendering and templating
    "jinja2>=3.1.6",
    "manim>=0.19.0",
    # Authentication and security
    "python-jose[cryptography]>=3.3.0", # JWT verification
    "httpx>=0.27.0", # HTTP client for Auth0 JWKS
    # Database
    "sqlalchemy>=2.0.0", # ORM
    "asyncpg>=0.29.0", # PostgreSQL async driver
    "alembic>=1.13.0", # Database migrations
    # AWS integration
    "boto3>=1.34.0", # S3 for video storage
    # Utilities
    "python-multipart>=0.0.9", # Form data parsing
    "python-dotenv>=1.0.0", # Environment variable loading
    "pydantic>=2.0.0", # Data validation
    "pydantic-settings>=2.0.0", # Settings management
    "fastapi>=0.121.1",
    "sympy>=1.13",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
]
